test: update SignIn tests to match with the upgraded login
